requirements.txt
feedparser
requests
beautifulsoup4
playwright
discord-webhook
openai
python-dotenv